AWS tests update
